+++ pkg/disttask/framework/storage/BUILD.bazel
@@ -3,6 +3,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "storage",
     srcs = [
+        "history.go",
+        "nodes.go",
+        "subtask_state.go",
         "task_state.go",
         "task_table.go",
     ],
