+++ pkg/disttask/framework/storage/task_table.go
@@ -16,7 +16,6 @@ package storage
 
 import (
 	"context"
-	"fmt"
 	"strconv"
 	"strings"
 	"sync/atomic"
@@ -31,7 +30,6 @@ import (
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
 	"github.com/pingcap/tidb/pkg/util/chunk"
-	"github.com/pingcap/tidb/pkg/util/cpu"
 	"github.com/pingcap/tidb/pkg/util/intest"
 	"github.com/pingcap/tidb/pkg/util/logutil"
 	"github.com/pingcap/tidb/pkg/util/sqlescape"
@@ -44,8 +42,9 @@ const (
 	defaultSubtaskKeepDays = 14
 
 	basicTaskColumns = `id, task_key, type, state, step, priority, concurrency, create_time`
+	// TaskColumns is the columns for task.
 	// TODO: dispatcher_id will update to scheduler_id later
-	taskColumns = basicTaskColumns + `, start_time, state_update_time, meta, dispatcher_id, error`
+	TaskColumns = basicTaskColumns + `, start_time, state_update_time, meta, dispatcher_id, error`
 	// InsertTaskColumns is the columns used in insert task.
 	InsertTaskColumns   = `task_key, type, state, priority, concurrency, step, meta, create_time`
 	basicSubtaskColumns = `id, step, task_key, type, exec_id, state, concurrency, create_time, ordinal`
@@ -146,8 +145,8 @@ func row2TaskBasic(r chunk.Row) *proto.Task {
 	return task
 }
 
-// row2Task converts a row to a task.
-func row2Task(r chunk.Row) *proto.Task {
+// Row2Task converts a row to a task.
+func Row2Task(r chunk.Row) *proto.Task {
 	task := row2TaskBasic(r)
 	var startTime, updateTime time.Time
 	if !r.IsNull(8) {
@@ -266,20 +265,6 @@ func (mgr *TaskManager) CreateTaskWithSession(ctx context.Context, se sessionctx
 	return taskID, nil
 }
 
-// GetOneTask get a task from task table, it's used by scheduler only.
-func (mgr *TaskManager) GetOneTask(ctx context.Context) (task *proto.Task, err error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+taskColumns+" from mysql.tidb_global_task where state = %? limit 1", proto.TaskStatePending)
-	if err != nil {
-		return task, err
-	}
-
-	if len(rs) == 0 {
-		return nil, nil
-	}
-
-	return row2Task(rs[0]), nil
-}
-
 // GetTopUnfinishedTasks implements the scheduler.TaskManager interface.
 func (mgr *TaskManager) GetTopUnfinishedTasks(ctx context.Context) (task []*proto.Task, err error) {
 	rs, err := mgr.ExecuteSQLWithNewSession(ctx,
@@ -312,7 +297,7 @@ func (mgr *TaskManager) GetTasksInStates(ctx context.Context, states ...interfac
 	}
 
 	rs, err := mgr.ExecuteSQLWithNewSession(ctx,
-		"select "+taskColumns+" from mysql.tidb_global_task "+
+		"select "+TaskColumns+" from mysql.tidb_global_task "+
 			"where state in ("+strings.Repeat("%?,", len(states)-1)+"%?)"+
 			" order by priority asc, create_time asc, id asc", states...)
 	if err != nil {
@@ -320,31 +305,14 @@ func (mgr *TaskManager) GetTasksInStates(ctx context.Context, states ...interfac
 	}
 
 	for _, r := range rs {
-		task = append(task, row2Task(r))
-	}
-	return task, nil
-}
-
-// GetTasksFromHistoryInStates gets the tasks in history table in the states.
-func (mgr *TaskManager) GetTasksFromHistoryInStates(ctx context.Context, states ...interface{}) (task []*proto.Task, err error) {
-	if len(states) == 0 {
-		return task, nil
-	}
-
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+taskColumns+" from mysql.tidb_global_task_history where state in ("+strings.Repeat("%?,", len(states)-1)+"%?)", states...)
-	if err != nil {
-		return task, err
-	}
-
-	for _, r := range rs {
-		task = append(task, row2Task(r))
+		task = append(task, Row2Task(r))
 	}
 	return task, nil
 }
 
 // GetTaskByID gets the task by the task ID.
 func (mgr *TaskManager) GetTaskByID(ctx context.Context, taskID int64) (task *proto.Task, err error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+taskColumns+" from mysql.tidb_global_task where id = %?", taskID)
+	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+TaskColumns+" from mysql.tidb_global_task where id = %?", taskID)
 	if err != nil {
 		return task, err
 	}
@@ -352,13 +320,13 @@ func (mgr *TaskManager) GetTaskByID(ctx context.Context, taskID int64) (task *pr
 		return nil, ErrTaskNotFound
 	}
 
-	return row2Task(rs[0]), nil
+	return Row2Task(rs[0]), nil
 }
 
 // GetTaskByIDWithHistory gets the task by the task ID from both tidb_global_task and tidb_global_task_history.
 func (mgr *TaskManager) GetTaskByIDWithHistory(ctx context.Context, taskID int64) (task *proto.Task, err error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+taskColumns+" from mysql.tidb_global_task where id = %? "+
-		"union select "+taskColumns+" from mysql.tidb_global_task_history where id = %?", taskID, taskID)
+	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+TaskColumns+" from mysql.tidb_global_task where id = %? "+
+		"union select "+TaskColumns+" from mysql.tidb_global_task_history where id = %?", taskID, taskID)
 	if err != nil {
 		return task, err
 	}
@@ -366,12 +334,12 @@ func (mgr *TaskManager) GetTaskByIDWithHistory(ctx context.Context, taskID int64
 		return nil, ErrTaskNotFound
 	}
 
-	return row2Task(rs[0]), nil
+	return Row2Task(rs[0]), nil
 }
 
 // GetTaskByKey gets the task by the task key.
 func (mgr *TaskManager) GetTaskByKey(ctx context.Context, key string) (task *proto.Task, err error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+taskColumns+" from mysql.tidb_global_task where task_key = %?", key)
+	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+TaskColumns+" from mysql.tidb_global_task where task_key = %?", key)
 	if err != nil {
 		return task, err
 	}
@@ -379,13 +347,13 @@ func (mgr *TaskManager) GetTaskByKey(ctx context.Context, key string) (task *pro
 		return nil, ErrTaskNotFound
 	}
 
-	return row2Task(rs[0]), nil
+	return Row2Task(rs[0]), nil
 }
 
 // GetTaskByKeyWithHistory gets the task from history table by the task key.
 func (mgr *TaskManager) GetTaskByKeyWithHistory(ctx context.Context, key string) (task *proto.Task, err error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+taskColumns+" from mysql.tidb_global_task where task_key = %?"+
-		"union select "+taskColumns+" from mysql.tidb_global_task_history where task_key = %?", key, key)
+	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select "+TaskColumns+" from mysql.tidb_global_task where task_key = %?"+
+		"union select "+TaskColumns+" from mysql.tidb_global_task_history where task_key = %?", key, key)
 	if err != nil {
 		return task, err
 	}
@@ -393,35 +361,7 @@ func (mgr *TaskManager) GetTaskByKeyWithHistory(ctx context.Context, key string)
 		return nil, ErrTaskNotFound
 	}
 
-	return row2Task(rs[0]), nil
-}
-
-// GetUsedSlotsOnNodes implements the scheduler.TaskManager interface.
-func (mgr *TaskManager) GetUsedSlotsOnNodes(ctx context.Context) (map[string]int, error) {
-	// concurrency of subtasks of some step is the same, we use max(concurrency)
-	// to make group by works.
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, `
-		select
-			exec_id, sum(concurrency)
-		from (
-			select exec_id, task_key, max(concurrency) concurrency
-			from mysql.tidb_background_subtask
-			where state in (%?, %?)
-			group by exec_id, task_key
-		) a
-		group by exec_id`,
-		proto.SubtaskStatePending, proto.SubtaskStateRunning,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	slots := make(map[string]int, len(rs))
-	for _, r := range rs {
-		val, _ := r.GetMyDecimal(1).ToInt()
-		slots[r.GetString(0)] = int(val)
-	}
-	return slots, nil
+	return Row2Task(rs[0]), nil
 }
 
 // row2BasicSubTask converts a row to a subtask with basic info
@@ -480,26 +420,6 @@ func (mgr *TaskManager) GetSubtasksByExecIDAndStepAndStates(ctx context.Context,
 	rs, err := mgr.ExecuteSQLWithNewSession(ctx, `select `+SubtaskColumns+` from mysql.tidb_background_subtask
 		where exec_id = %? and task_key = %? and step = %?
 		and state in (`+strings.Repeat("%?,", len(states)-1)+"%?)", args...)
-	if err != nil {
-		return nil, err
-	}
-
-	subtasks := make([]*proto.Subtask, len(rs))
-	for i, row := range rs {
-		subtasks[i] = Row2SubTask(row)
-	}
-	return subtasks, nil
-}
-
-// GetSubtasksByExecIdsAndStepAndState gets all subtasks by given taskID, exec_id, step and state.
-func (mgr *TaskManager) GetSubtasksByExecIdsAndStepAndState(ctx context.Context, execIDs []string, taskID int64, step proto.Step, state proto.SubtaskState) ([]*proto.Subtask, error) {
-	args := []interface{}{taskID, step, state}
-	for _, execID := range execIDs {
-		args = append(args, execID)
-	}
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, `select `+SubtaskColumns+` from mysql.tidb_background_subtask
-		where task_key = %? and step = %? and state = %?
-		and exec_id in (`+strings.Repeat("%?,", len(execIDs)-1)+"%?)", args...)
 	if err != nil {
 		return nil, err
 	}
@@ -571,7 +491,7 @@ func (mgr *TaskManager) GetActiveSubtasks(ctx context.Context, taskID int64) ([]
 	return subtasks, nil
 }
 
-// GetSubtasksByStepAndState gets the subtask by step and state.
+// GetAllSubtasksByStepAndState gets the subtask by step and state.
 func (mgr *TaskManager) GetAllSubtasksByStepAndState(ctx context.Context, taskID int64, step proto.Step, state proto.SubtaskState) ([]*proto.Subtask, error) {
 	rs, err := mgr.ExecuteSQLWithNewSession(ctx, `select `+SubtaskColumns+` from mysql.tidb_background_subtask
 		where task_key = %? and state = %? and step = %?`,
@@ -634,8 +554,8 @@ func (mgr *TaskManager) GetSubtaskCntGroupByStates(ctx context.Context, taskID i
 	return res, nil
 }
 
-// CollectSubTaskError collects the subtask error.
-func (mgr *TaskManager) CollectSubTaskError(ctx context.Context, taskID int64) ([]error, error) {
+// GetSubtaskErrors gets subtasks' errors.
+func (mgr *TaskManager) GetSubtaskErrors(ctx context.Context, taskID int64) ([]error, error) {
 	rs, err := mgr.ExecuteSQLWithNewSession(ctx,
 		`select error from mysql.tidb_background_subtask
              where task_key = %? AND state in (%?, %?)`, taskID, proto.SubtaskStateFailed, proto.SubtaskStateCanceled)
@@ -680,64 +600,6 @@ func (mgr *TaskManager) HasSubtasksInStates(ctx context.Context, tidbID string, 
 	return len(rs) > 0, nil
 }
 
-// StartSubtask updates the subtask state to running.
-func (mgr *TaskManager) StartSubtask(ctx context.Context, subtaskID int64, execID string) error {
-	err := mgr.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		vars := se.GetSessionVars()
-		_, err := sqlexec.ExecSQL(ctx,
-			se,
-			`update mysql.tidb_background_subtask
-			 set state = %?, start_time = unix_timestamp(), state_update_time = unix_timestamp()
-			 where id = %? and exec_id = %?`,
-			proto.SubtaskStateRunning,
-			subtaskID,
-			execID)
-		if err != nil {
-			return err
-		}
-		if vars.StmtCtx.AffectedRows() == 0 {
-			return ErrSubtaskNotFound
-		}
-		return nil
-	})
-	return err
-}
-
-// InitMeta insert the manager information into dist_framework_meta.
-func (mgr *TaskManager) InitMeta(ctx context.Context, tidbID string, role string) error {
-	return mgr.WithNewSession(func(se sessionctx.Context) error {
-		return mgr.InitMetaSession(ctx, se, tidbID, role)
-	})
-}
-
-// InitMetaSession insert the manager information into dist_framework_meta.
-// if the record exists, update the cpu_count and role.
-func (*TaskManager) InitMetaSession(ctx context.Context, se sessionctx.Context, execID string, role string) error {
-	cpuCount := cpu.GetCPUCount()
-	_, err := sqlexec.ExecSQL(ctx, se, `
-		insert into mysql.dist_framework_meta(host, role, cpu_count, keyspace_id)
-		values (%?, %?, %?, -1)
-		on duplicate key
-		update cpu_count = %?, role = %?`,
-		execID, role, cpuCount, cpuCount, role)
-	return err
-}
-
-// RecoverMeta insert the manager information into dist_framework_meta.
-// if the record exists, update the cpu_count.
-// Don't update role for we only update it in `set global tidb_service_scope`.
-// if not there might has a data race.
-func (mgr *TaskManager) RecoverMeta(ctx context.Context, execID string, role string) error {
-	cpuCount := cpu.GetCPUCount()
-	_, err := mgr.ExecuteSQLWithNewSession(ctx, `
-		insert into mysql.dist_framework_meta(host, role, cpu_count, keyspace_id)
-		values (%?, %?, %?, -1)
-		on duplicate key
-		update cpu_count = %?`,
-		execID, role, cpuCount, cpuCount)
-	return err
-}
-
 // UpdateSubtaskStateAndError updates the subtask state.
 func (mgr *TaskManager) UpdateSubtaskStateAndError(
 	ctx context.Context,
@@ -749,50 +611,10 @@ func (mgr *TaskManager) UpdateSubtaskStateAndError(
 	return err
 }
 
-// FinishSubtask updates the subtask meta and mark state to succeed.
-func (mgr *TaskManager) FinishSubtask(ctx context.Context, execID string, id int64, meta []byte) error {
-	_, err := mgr.ExecuteSQLWithNewSession(ctx, `update mysql.tidb_background_subtask
-		set meta = %?, state = %?, state_update_time = unix_timestamp(), end_time = CURRENT_TIMESTAMP()
-		where id = %? and exec_id = %?`,
-		meta, proto.SubtaskStateSucceed, id, execID)
-	return err
-}
-
-// DeleteSubtasksByTaskID deletes the subtask of the given task ID.
-func (mgr *TaskManager) DeleteSubtasksByTaskID(ctx context.Context, taskID int64) error {
-	_, err := mgr.ExecuteSQLWithNewSession(ctx, `delete from mysql.tidb_background_subtask
-		where task_key = %?`, taskID)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
 // GetTaskExecutorIDsByTaskID gets the task executor IDs of the given task ID.
 func (mgr *TaskManager) GetTaskExecutorIDsByTaskID(ctx context.Context, taskID int64) ([]string, error) {
 	rs, err := mgr.ExecuteSQLWithNewSession(ctx, `select distinct(exec_id) from mysql.tidb_background_subtask
 		where task_key = %?`, taskID)
-	if err != nil {
-		return nil, err
-	}
-	if len(rs) == 0 {
-		return nil, nil
-	}
-
-	instanceIDs := make([]string, 0, len(rs))
-	for _, r := range rs {
-		id := r.GetString(0)
-		instanceIDs = append(instanceIDs, id)
-	}
-
-	return instanceIDs, nil
-}
-
-// GetTaskExecutorIDsByTaskIDAndStep gets the task executor IDs of the given global task ID and step.
-func (mgr *TaskManager) GetTaskExecutorIDsByTaskIDAndStep(ctx context.Context, taskID int64, step proto.Step) ([]string, error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, `select distinct(exec_id) from mysql.tidb_background_subtask
-		where task_key = %? and step = %?`, taskID, step)
 	if err != nil {
 		return nil, err
 	}
@@ -828,64 +650,6 @@ func (mgr *TaskManager) UpdateSubtasksExecIDs(ctx context.Context, subtasks []*p
 		}
 		return nil
 	})
-	return err
-}
-
-// RunningSubtasksBack2Pending implements the taskexecutor.TaskTable interface.
-func (mgr *TaskManager) RunningSubtasksBack2Pending(ctx context.Context, subtasks []*proto.Subtask) error {
-	// skip the update process.
-	if len(subtasks) == 0 {
-		return nil
-	}
-	err := mgr.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		for _, subtask := range subtasks {
-			_, err := sqlexec.ExecSQL(ctx, se, `
-				update mysql.tidb_background_subtask
-				set state = %?, state_update_time = CURRENT_TIMESTAMP()
-				where id = %? and exec_id = %? and state = %?`,
-				proto.SubtaskStatePending, subtask.ID, subtask.ExecID, proto.SubtaskStateRunning)
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	return err
-}
-
-// DeleteDeadNodes deletes the dead nodes from mysql.dist_framework_meta.
-func (mgr *TaskManager) DeleteDeadNodes(ctx context.Context, nodes []string) error {
-	if len(nodes) == 0 {
-		return nil
-	}
-	return mgr.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		deleteSQL := new(strings.Builder)
-		if err := sqlescape.FormatSQL(deleteSQL, "delete from mysql.dist_framework_meta where host in("); err != nil {
-			return err
-		}
-		deleteElems := make([]string, 0, len(nodes))
-		for _, node := range nodes {
-			deleteElems = append(deleteElems, fmt.Sprintf(`"%s"`, node))
-		}
-
-		deleteSQL.WriteString(strings.Join(deleteElems, ", "))
-		deleteSQL.WriteString(")")
-		_, err := sqlexec.ExecSQL(ctx, se, deleteSQL.String())
-		return err
-	})
-}
-
-// PauseSubtasks update all running/pending subtasks to pasued state.
-func (mgr *TaskManager) PauseSubtasks(ctx context.Context, execID string, taskID int64) error {
-	_, err := mgr.ExecuteSQLWithNewSession(ctx,
-		`update mysql.tidb_background_subtask set state = "paused" where task_key = %? and state in ("running", "pending") and exec_id = %?`, taskID, execID)
-	return err
-}
-
-// ResumeSubtasks update all paused subtasks to pending state.
-func (mgr *TaskManager) ResumeSubtasks(ctx context.Context, taskID int64) error {
-	_, err := mgr.ExecuteSQLWithNewSession(ctx,
-		`update mysql.tidb_background_subtask set state = "pending", error = null where task_key = %? and state = "paused"`, taskID)
 	return err
 }
 
@@ -1115,19 +879,6 @@ func serializeErr(err error) []byte {
 	return errBytes
 }
 
-// IsTaskCancelling checks whether the task state is cancelling.
-func (mgr *TaskManager) IsTaskCancelling(ctx context.Context, taskID int64) (bool, error) {
-	rs, err := mgr.ExecuteSQLWithNewSession(ctx, "select 1 from mysql.tidb_global_task where id=%? and state = %?",
-		taskID, proto.TaskStateCancelling,
-	)
-
-	if err != nil {
-		return false, err
-	}
-
-	return len(rs) > 0, nil
-}
-
 // GetSubtasksWithHistory gets the subtasks from tidb_global_task and tidb_global_task_history.
 func (mgr *TaskManager) GetSubtasksWithHistory(ctx context.Context, taskID int64, step proto.Step) ([]*proto.Subtask, error) {
 	var (
@@ -1168,162 +919,4 @@ func (mgr *TaskManager) GetSubtasksWithHistory(ctx context.Context, taskID int64
 		subtasks = append(subtasks, Row2SubTask(r))
 	}
 	return subtasks, nil
-}
-
-// TransferSubtasks2HistoryWithSession transfer the selected subtasks into tidb_background_subtask_history table by taskID.
-func (*TaskManager) TransferSubtasks2HistoryWithSession(ctx context.Context, se sessionctx.Context, taskID int64) error {
-	_, err := sqlexec.ExecSQL(ctx, se, `insert into mysql.tidb_background_subtask_history select * from mysql.tidb_background_subtask where task_key = %?`, taskID)
-	if err != nil {
-		return err
-	}
-	// delete taskID subtask
-	_, err = sqlexec.ExecSQL(ctx, se, "delete from mysql.tidb_background_subtask where task_key = %?", taskID)
-	return err
-}
-
-// TransferTasks2History transfer the selected tasks into tidb_global_task_history table by taskIDs.
-func (mgr *TaskManager) TransferTasks2History(ctx context.Context, tasks []*proto.Task) error {
-	if len(tasks) == 0 {
-		return nil
-	}
-	taskIDStrs := make([]string, 0, len(tasks))
-	for _, task := range tasks {
-		taskIDStrs = append(taskIDStrs, fmt.Sprintf("%d", task.ID))
-	}
-	return mgr.WithNewTxn(ctx, func(se sessionctx.Context) error {
-		// sensitive data in meta might be redacted, need update first.
-		for _, t := range tasks {
-			_, err := sqlexec.ExecSQL(ctx, se, `
-				update mysql.tidb_global_task
-				set meta= %?, state_update_time = CURRENT_TIMESTAMP()
-				where id = %?`, t.Meta, t.ID)
-			if err != nil {
-				return err
-			}
-		}
-		_, err := sqlexec.ExecSQL(ctx, se, `
-			insert into mysql.tidb_global_task_history
-			select * from mysql.tidb_global_task
-			where id in(`+strings.Join(taskIDStrs, `, `)+`)`)
-		if err != nil {
-			return err
-		}
-
-		_, err = sqlexec.ExecSQL(ctx, se, `
-			delete from mysql.tidb_global_task
-			where id in(`+strings.Join(taskIDStrs, `, `)+`)`)
-
-		for _, t := range tasks {
-			err = mgr.TransferSubtasks2HistoryWithSession(ctx, se, t.ID)
-			if err != nil {
-				return err
-			}
-		}
-		return err
-	})
-}
-
-// GCSubtasks deletes the history subtask which is older than the given days.
-func (mgr *TaskManager) GCSubtasks(ctx context.Context) error {
-	subtaskHistoryKeepSeconds := defaultSubtaskKeepDays * 24 * 60 * 60
-	failpoint.Inject("subtaskHistoryKeepSeconds", func(val failpoint.Value) {
-		if val, ok := val.(int); ok {
-			subtaskHistoryKeepSeconds = val
-		}
-	})
-	_, err := mgr.ExecuteSQLWithNewSession(
-		ctx,
-		fmt.Sprintf("DELETE FROM mysql.tidb_background_subtask_history WHERE state_update_time < UNIX_TIMESTAMP() - %d ;", subtaskHistoryKeepSeconds),
-	)
-	return err
-}
-
-// GetManagedNodes implements scheduler.TaskManager interface.
-func (mgr *TaskManager) GetManagedNodes(ctx context.Context) ([]proto.ManagedNode, error) {
-	var nodes []proto.ManagedNode
-	err := mgr.WithNewSession(func(se sessionctx.Context) error {
-		var err2 error
-		nodes, err2 = mgr.getManagedNodesWithSession(ctx, se)
-		return err2
-	})
-	return nodes, err
-}
-
-func (mgr *TaskManager) getManagedNodesWithSession(ctx context.Context, se sessionctx.Context) ([]proto.ManagedNode, error) {
-	nodes, err := mgr.getAllNodesWithSession(ctx, se)
-	if err != nil {
-		return nil, err
-	}
-	nodeMap := make(map[string][]proto.ManagedNode, 2)
-	for _, node := range nodes {
-		nodeMap[node.Role] = append(nodeMap[node.Role], node)
-	}
-	if len(nodeMap["background"]) == 0 {
-		return nodeMap[""], nil
-	}
-	return nodeMap["background"], nil
-}
-
-// GetAllNodes gets nodes in dist_framework_meta.
-func (mgr *TaskManager) GetAllNodes(ctx context.Context) ([]proto.ManagedNode, error) {
-	var nodes []proto.ManagedNode
-	err := mgr.WithNewSession(func(se sessionctx.Context) error {
-		var err2 error
-		nodes, err2 = mgr.getAllNodesWithSession(ctx, se)
-		return err2
-	})
-	return nodes, err
-}
-
-func (*TaskManager) getAllNodesWithSession(ctx context.Context, se sessionctx.Context) ([]proto.ManagedNode, error) {
-	rs, err := sqlexec.ExecSQL(ctx, se, `
-		select host, role, cpu_count
-		from mysql.dist_framework_meta
-		order by host`)
-	if err != nil {
-		return nil, err
-	}
-	nodes := make([]proto.ManagedNode, 0, len(rs))
-	for _, r := range rs {
-		nodes = append(nodes, proto.ManagedNode{
-			ID:       r.GetString(0),
-			Role:     r.GetString(1),
-			CPUCount: int(r.GetInt64(2)),
-		})
-	}
-	return nodes, nil
-}
-
-// GetCPUCountOfManagedNode gets the cpu count of managed node.
-func (mgr *TaskManager) GetCPUCountOfManagedNode(ctx context.Context) (int, error) {
-	var cnt int
-	err := mgr.WithNewSession(func(se sessionctx.Context) error {
-		var err2 error
-		cnt, err2 = mgr.getCPUCountOfManagedNode(ctx, se)
-		return err2
-	})
-	return cnt, err
-}
-
-// getCPUCountOfManagedNode gets the cpu count of managed node.
-// returns error when there's no managed node or no node has valid cpu count.
-func (mgr *TaskManager) getCPUCountOfManagedNode(ctx context.Context, se sessionctx.Context) (int, error) {
-	nodes, err := mgr.getManagedNodesWithSession(ctx, se)
-	if err != nil {
-		return 0, err
-	}
-	if len(nodes) == 0 {
-		return 0, errors.New("no managed nodes")
-	}
-	var cpuCount int
-	for _, n := range nodes {
-		if n.CPUCount > 0 {
-			cpuCount = n.CPUCount
-			break
-		}
-	}
-	if cpuCount == 0 {
-		return 0, errors.New("no managed node have enough resource for dist task")
-	}
-	return cpuCount, nil
 }
